made it so that the documentaiton page runs - current issue: FullCalendar in the JS file
